+++ mos/src/debugger/adapters/test_runner/mod.rs
@@ -296,13 +296,13 @@ impl MachineAdapter for TestRunnerAdapter {
         Ok(self.runner.read().unwrap().cpu().get_status_register())
     }
 
-    fn set_variable(&mut self, name: String, value: i64) -> MosResult<()> {
+    fn set_variable(&mut self, name: String, value: u8) -> MosResult<()> {
         let mut write_runner = self.runner.write().unwrap();
         let cpu_mut = write_runner.cpu_mut();
         match name.as_str() {
-            "A" => cpu_mut.set_accumulator(value as u8),
-            "X" => cpu_mut.set_x_register(value as u8),
-            "Y" => cpu_mut.set_y_register(value as u8),
+            "A" => cpu_mut.set_accumulator(value),
+            "X" => cpu_mut.set_x_register(value),
+            "Y" => cpu_mut.set_y_register(value),
             _ => {}
         }
